% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.choose_word}
\alias{.choose_word}
\title{Choose a word from a dictionnary}
\usage{
.choose_word(dict, n.letters)
}
\arguments{
\item{dict}{A vetor of words to choose from}

\item{n.letters}{The number of letters of the word to be chosen}
}
\value{
A word. Wil throw an error if no word of the wanted length exists in
the dictionnary
}
\description{
Choose a word of the given length from a dictionnary
}
