% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.display_letters}
\alias{.display_letters}
\title{Display said letters on the console}
\usage{
.display_letters(word, said.letters)
}
\arguments{
\item{word}{The word}

\item{said.letters}{Letters that have already been said}
}
\value{
nothing
}
\description{
Display the letters that have been said, but not letters that are in the word
}
