% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.display_word}
\alias{.display_word}
\title{Display the word on the console}
\usage{
.display_word(word, found.letters)
}
\arguments{
\item{word}{The word}

\item{found.letters}{Letters that have already been found}
}
\value{
nothing
}
\description{
Display the word on the console, with only found letters shown
}
