% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hangman.R
\name{hangman}
\alias{hangman}
\title{Play the hangman game}
\usage{
hangman(dictionnary = "french", n.letters = 6, graph = TRUE)
}
\arguments{
\item{dictionnary}{Either one of 'french' or 'english'. The dictionnary from
which to select a word}

\item{n.letters}{The number of letters of the wanted word}

\item{graph}{Display or not the awesome graphics, default to TRUE. Will open
a graphical window.}
}
\value{
nothing
}
\description{
Sets a game of hangman with a chosen word or a randomly selected word
}
